Moved usings inside the namespace. Moved DataBase,DataSerialization into Helpers/Data namespace. Add ScoreBoardDatabase class - needs implementation. Removed unnecessary comments.
