+++ Hangman/Hangman-4/Helpers/Data/DataSerialization.cs
@@ -1,11 +1,11 @@
-﻿namespace HangMan.InputProviders
+﻿namespace HangMan.InputProviders.Data
 {
     using HangMan.GameObjects;
-using HangMan.Interfaces;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.IO;
+    using HangMan.Interfaces;
+    using System;
+    using System.Collections;
+    using System.Collections.Generic;
+    using System.IO;
 
     public class DataSerialization : IDataSerialization
     {
