+++ Hangman/Hangman-4/Helpers/Data/ScoreBoardDatabase.cs
@@ -1,0 +1,21 @@
+﻿using HangMan.Interfaces;
+
+namespace HangMan.Helpers.Data
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Text;
+
+    public class ScoreBoardDatabase : DataBase
+    {
+        public ScoreBoardDatabase(IDataSerialization dataSerialization)
+            :base(dataSerialization)
+        {
+        }
+
+        //Read scoreboard
+
+        //Write scoreboard
+    }
+}
